requirements.txt
django
django-bootstrap4
django-rest-framework
gunicorn
psycopg2-binary